Heath Drop
I tried to set up the health drop and for that I had to create a new
sprite. Don't know why it's not work, I took the ammo_drop code so it
would be identical but it's not work. So i would like to know where I
screwed up. thankx